requirements.txt
requests==2.18.4
pyyaml==5.3.1
